requirements.txt
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
Django==4.2.7
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
idna==3.8
psycopg2-binary==2.9.9
PyJWT==2.9.0
python-decouple==3.8
requests==2.32.3
sqlparse==0.5.1
urllib3==2.2.2